Fix warning.

+++ src/drivers/Unix/Fl_Unix_System_Driver.cxx
@@ -544,7 +544,7 @@ char *Fl_Unix_System_Driver::preference_user_rootnode(
       prefs_path_14.append('/');
     if (prefs_path_14.find("~/")==0) // starts with "~" 
       prefs_path_14.replace(0, 1, home_path);
-    size_t h_env = prefs_path_14.find("${HOME}");
+    int h_env = prefs_path_14.find("${HOME}");
     if (h_env!=prefs_path_14.npos)
       prefs_path_14.replace(h_env, 7, home_path);
     h_env = prefs_path_14.find("$HOME/");
